Update creating-a-github-app-from-a-manifest.md
I've raised a PR to add some extra info on the manifest parameters. 

The "github parameters" method for creating an app details many more parameters that can be set, but the manifest only suggests these to be valid. I actually tried setting the `callback_url` and the `request_oauth_on_install` and that actually worked.

`redirect_url` also suggests it is the callback url, but it is not, it's where the user gets redirected to after creating the app.